Добавляем мемоизацию на дочерние компоненты страниц.
