+++ src/components/footer/footer.tsx
@@ -1,5 +1,6 @@
 import {Link} from 'react-router-dom';
 import { AppRoute } from '../../const';
+import { memo } from 'react';
 
 function Footer(): JSX.Element {
   return (
@@ -11,4 +12,4 @@ function Footer(): JSX.Element {
   );
 }
 
-export default Footer;
+export const MemoizedFooter = memo(Footer);
