+++ src/pages/favorites-screen/favorites-screen.tsx
@@ -1,4 +1,4 @@
-import FavoritesLocations from './favorites-screen-components/favorites-locations';
+import { MemoizedFavoritesLocations } from './favorites-screen-components/favorites-locations';
 import { OffersProps } from '../../types/offer';
 import { AuthorizationStatus } from '../../const';
 
@@ -15,7 +15,7 @@ function FavoritesScreen({offers, authorizationStatus}: FavoritesScreenProps): J
         <section className={favoriteClassName}>
           <h1 className={`${favoriteClassName}__title`}>Saved listing</h1>
           <ul className={`${favoriteClassName}__list`}>
-            <FavoritesLocations offers={offers} favoritesClassName={favoriteClassName} authorizationStatus={authorizationStatus} />
+            <MemoizedFavoritesLocations offers={offers} favoritesClassName={favoriteClassName} authorizationStatus={authorizationStatus} />
           </ul>
         </section>
       </div>
