+++ src/pages/offer-screen/offer-screen.tsx
@@ -1,12 +1,12 @@
-import OfferGallery from './offer-screen-components/offer-gallery';
-import OfferComponent from './offer-screen-components/offer-component';
-import OfferReviewList from './offer-screen-components/offer-review-list';
-import OfferReviewForm from './offer-screen-components/offer-review-form';
-import OfferCard from '../../components/card/card';
+import { MemoizedOfferGallery } from './offer-screen-components/offer-gallery';
+import { MemoizedOfferComponent } from './offer-screen-components/offer-component';
+import { MemoizedOfferReviewList } from './offer-screen-components/offer-review-list';
+import { MemoizedOfferReviewForm } from './offer-screen-components/offer-review-form';
+import { MemoizedOfferCard } from '../../components/card/card';
 import { CardProps } from '../../types/offer';
 import NotFoundScreen from '../not-found-screen/not-found-screen';
 import { AuthorizationStatus } from '../../const';
-import Map from '../../components/map/map';
+import { MemoizedMap } from '../../components/map/map';
 import { ReviewsProps } from '../../types/review';
 import { fetchCurrentOffer, fetchNearOffers, fetchOfferReviews } from '../../store/api-actions';
 import { useAppDispatch, useAppSelector } from '../../hooks';
@@ -54,12 +54,12 @@ function OfferScreen({authorizationStatus}: OfferScreenProps): JSX.Element {
     <main className={`page__main page__main--${offerPageClassName}`}>
       <section className={offerPageClassName}>
         {currentOffer.images ?
-          <OfferGallery
+          <MemoizedOfferGallery
             images={currentOffer.images}
           /> : null}
         <div className={`${offerPageClassName}__container container`}>
           <div className={`${offerPageClassName}__wrapper`}>
-            <OfferComponent
+            <MemoizedOfferComponent
               key={currentOffer.id}
               offer={currentOffer}
               offerClassName={offerPageClassName}
@@ -70,7 +70,7 @@ function OfferScreen({authorizationStatus}: OfferScreenProps): JSX.Element {
                 (
                   <ul className="reviews__list">
                     {offerReviews.map((review: ReviewsProps) => (
-                      <OfferReviewList
+                      <MemoizedOfferReviewList
                         key={review.id}
                         offerReview={review}
                       />
@@ -79,7 +79,7 @@ function OfferScreen({authorizationStatus}: OfferScreenProps): JSX.Element {
                 ) : <p>This offer do not have any reviews.</p>}
               {
                 authorizationStatus === AuthorizationStatus.Auth ?
-                  <OfferReviewForm id={id} />
+                  <MemoizedOfferReviewForm id={id} />
                   : <b>Only authorized user could leave a review. Please Sign in</b>
               }
             </section>
@@ -87,7 +87,7 @@ function OfferScreen({authorizationStatus}: OfferScreenProps): JSX.Element {
         </div>
         {nearOffers ?
           (
-            <Map city={currentOffer.city} offers={nearOffers} activeOfferId={currentOffer.id} mapClassName={offerPageClassName} />
+            <MemoizedMap city={currentOffer.city} offers={nearOffers} activeOfferId={currentOffer.id} mapClassName={offerPageClassName} />
           ) : null}
       </section>
       <div className="container">
@@ -97,7 +97,7 @@ function OfferScreen({authorizationStatus}: OfferScreenProps): JSX.Element {
               <h2 className="near-places__title">Other places in the neighbourhood</h2>
               <div className="near-places__list places__list">
                 {nearOffers.map((offer: CardProps) => (
-                  <OfferCard
+                  <MemoizedOfferCard
                     key={offer.id}
                     offer={offer}
                     cardClassName={nearPlacesClassName}
